requirements.txt

streamlit==1.37.1
feedparser==6.0.11
requests==2.32.3
readability-lxml==0.8.1
lxml==5.2.2
beautifulsoup4==4.12.3
spacy==3.7.5
dateparser==1.2.0
python-dateutil==2.9.0.post0
rapidfuzz==3.9.7
newspaper3k==0.2.8
